Add files via upload
The heart disease prediction using logistic regression is a project designed to predict the likelihood of a person having heart disease based on various medical and demographic factors. Logistic regression is a statistical method used for binary classification, making it suitable for predicting whether a person has heart disease or not.